+++ App/Resource/Translations/RabbitCommonApp_pt_BR.ts
@@ -74,12 +74,12 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="129"/>
+        <location filename="../../MainWindow.cpp" line="130"/>
         <source>Encrypt:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="129"/>
+        <location filename="../../MainWindow.cpp" line="130"/>
         <source>Source:</source>
         <translation type="unfinished"></translation>
     </message>